// WebClient/src/app/models/ISale.interface.ts
export interface ISale {
	Id?: Number;
	SaleDate?: Date;
	Code?: String;
	CreatedBy?: String;
	UpdatedBy?: String;
	CustomerId?: Number;
	CustomerName?: String;
	CustomerAddress?: String;
	CustomerPhone?: String;
	ProductCode?: String;
	ProductId?: Number;
	ProductName?: String;
	CategoryName?: String;
	UnitCode?: String;
	UnitId?: Number;
	ProductForm?: String;
	Production?: String;
	ProductPrice?: Number;
	Quantity?: Number;
	Total?: Number;
	Paid?: Number;
	Note?: String;
}